requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
langgraph
pydantic
python-multipart
python-dotenv
chromadb
sentence-transformers
pypdf
tavily-python
requests
numpy
tiktoken